requirements.txt
six~=1.16.0
# torch~=2.0.0
gensim~=4.3.2
tqdm~=4.65.0
dgl~=1.1.2
scikit-learn~=1.3.0
mknapsack~=1.1.12
capstone~=4.0.2
networkx~=3.0
scipy~=1.10.1
matplotlib~=3.7.1
transformers==4.32
cython~=0.29.34
pandas
pyelftools~=0.29
psutil~=5.9.4
filelock~=3.12.2
pybind11~=2.11.1
PyYAML~=6.0.1
rocksdb
tensorflow==2.8.0
bert_pytorch
pandas==1.4.4
accelerate
mambda-ssm


